Move CMC icon and setup document to common `ConnectMyComputer` directory
