+++ web/packages/teleterm/src/ui/ConnectMyComputer/DocumentSetup/DocumentSetup.story.tsx
@@ -1,11 +1,11 @@
 /**
- * Copyright 2023 Gravitational, Inc
+ * Copyright 2023 Gravitational, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,4 +14,14 @@
  * limitations under the License.
  */
 
-export { ConnectMyComputer } from './ConnectMyComputer';
+import React from 'react';
+
+import { DocumentSetup } from './DocumentSetup';
+
+export default {
+  title: 'Teleterm/ConnectMyComputer/DocumentSetup',
+};
+
+export function Default() {
+  return <DocumentSetup visible={true} doc={undefined} />;
+}
