+++ web/packages/teleterm/src/ui/DocumentCluster/ClusterResources/ClusterResources.tsx
@@ -19,6 +19,7 @@ import styled from 'styled-components';
 import { Flex } from 'design';
 
 import { useClusterContext } from 'teleterm/ui/DocumentCluster/clusterContext';
+import { ConnectMyComputerNavigationMenu } from 'teleterm/ui/ConnectMyComputer';
 
 import SideNav from './SideNav';
 import Servers from './Servers';
@@ -41,7 +42,10 @@ export default function ClusterResources() {
   return (
     <StyledMain>
       <Flex pb={5} flexDirection="column">
-        <SideNav mb={2} />
+        <Flex justifyContent="space-between">
+          <SideNav mb={2} />
+          <ConnectMyComputerNavigationMenu />
+        </Flex>
         <HorizontalSplit>
           {clusterCtx.isLocationActive('/resources/servers') && <Servers />}
           {clusterCtx.isLocationActive('/resources/databases') && <Databases />}
