+++ web/packages/teleterm/src/ui/DocumentCluster/ClusterResources/SideNav/SideNav.tsx
@@ -51,7 +51,6 @@ export type SideNavItem = {
 const StyledNav = styled(Flex)`
   min-width: 180px;
   overflow: auto;
-  height: 100%;
 `;
 
 function createItems(): SideNavItem[] {
